requirements.txt
fastmcp==0.5.0
mcp[cli]
selenium==4.15.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
langchain-google-genai==1.0.8
langchain-core==0.3.0
streamlit==1.28.0